+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - added `dateIsValid`, `dateIsLastDayOfMonth` and `dateDifferenceInDays` utility functions
+- added `newGuid` function and `emptyGuid` constant
 
 ## [0.1.1] - 2023-08-04